Share OS Doppler secret loader
